+++ lib/wasi-types-generated/src/lib.rs
@@ -1,3 +1,5 @@
+pub mod wasi;
 pub mod wasi_filesystem;
 pub mod wasi_io_typenames;
-pub mod wasi_snapshot0;
+
+pub mod wasi_snapshot0 {}
